% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urls.R
\name{get_base_url_}
\alias{get_base_url_}
\title{Get the base URL of GHSL's download server}
\usage{
get_base_url_()
}
\value{
GHSL's download server base URL
}
\description{
Get the base URL of GHSL's download server
}
