% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urls.R
\name{get_product_url}
\alias{get_product_url}
\title{Get the download server URL of the chosen dataset, product and tile}
\usage{
get_product_url(dataset, product, row, col)
}
\arguments{
\item{dataset}{String - GHSL dataset name}

\item{product}{String - product name within chosen dataset}

\item{row}{Numeric - tile row number}

\item{col}{Numeric - tile column number}
}
\value{
String - GHSL download server URL of chosen dataset, product and tile
}
\description{
Get the download server URL of the chosen dataset, product and tile
}
\examples{
get_product_url("GHS-BUILT-S", "GHS_BUILT_S_E2018_GLOBE_R2022A_54009_10", 1, 1)
}
