% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghsl_dl.R
\name{get_row_col_}
\alias{get_row_col_}
\title{Helper function, convert inputted sfc_POINT to GHSL's tile row and column
NB: This function is not intended for users}
\usage{
get_row_col_(point)
}
\arguments{
\item{point}{An \code{sfc_POINT} object}
}
\value{
A row and column vector
}
\description{
Helper function, convert inputted sfc_POINT to GHSL's tile row and column
NB: This function is not intended for users
}
