% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghslDL-package.R
\docType{package}
\name{ghslDL-package}
\alias{ghslDL}
\alias{ghslDL-package}
\title{ghslDL: Downloader for GHSL datasets}
\description{
Downloader for GHSL datasets and products. Requires spatial point for tile-based download.
}
\keyword{internal}
