% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List available and supported datasets for download from GHSL}
\usage{
list_datasets()
}
\value{
Dataframe - supported datasets with names and descriptions
}
\description{
List available and supported datasets for download from GHSL
}
\examples{
list_datasets()
}
