% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_products.R
\name{list_products}
\alias{list_products}
\title{List available products within chosen dataset}
\usage{
list_products(dataset, cache = TRUE)
}
\arguments{
\item{dataset}{String - GHSL dataset name}

\item{cache}{Logical - default is \code{TRUE}, enables HTTP request caching with
\code{httpcache}. If \code{FALSE}, uncached request.}
}
\value{
Dataframe - products available for download within chosen dataset
}
\description{
List available products within chosen dataset
}
\examples{
list_products("GHS-BUILT-V")
}
