% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghsl_dl.R
\name{ghsl_dl}
\alias{ghsl_dl}
\title{GHSL downloader, helps download raster with specified dataset, product and
spatial point}
\usage{
ghsl_dl(dataset, point, overwrite = FALSE, cache = TRUE)
}
\arguments{
\item{dataset}{String - dataset name}

\item{point}{An \code{sf} point}

\item{overwrite}{Logical - default is \code{FALSE}, if \code{TRUE} allows the
downloader to overwrite existing downloaded files}

\item{cache}{Logical - default is \code{TRUE}, enables HTTP request caching with
\code{httpcache}. If \code{FALSE}, uncached request.}
}
\value{
The download server response after downloaded
}
\description{
GHSL downloader, helps download raster with specified dataset, product and
spatial point
}
\examples{
\dontrun{
ghsl_dl("GHS-BUILT-S", sf::st_sfc(sf::st_point(c(106.6297, 10.8231)), crs=4326))
}
}
